requirements.txt
# SEARS Voice AI Diagnostic Agent - Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Pydantic
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# HTTP & WebSockets
httpx>=0.26.0
websockets>=12.0
aiohttp>=3.9.0

# OpenAI
openai>=1.12.0

# Twilio
twilio>=8.13.0

# Email
sendgrid>=6.11.0

# Image Processing
pillow>=10.2.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
